Shutdown EventsLogger cleanly, process all accumulated items in the queue each tick (#22411)
The EventLogger previously would infinite loop when all senders went away. Now it shuts down cleanly. Waiting for all senders to go away is a pretty safe/convenient mechanism to let us shut down when we know no more records can be sent.

I've also used `try_recv` on the channel in a loop up to our maximum buffer size each time something is put into the channel. In the benchmark this makes a noticeable difference in throughput (~2x). This does increase the risk that we'll exceed the maximum put record batch size (500 records, 1000 kb / record, 4MB total). I'll check our monitoring and see if we should do anything about this in the near term.

I've also added a small benchmark for this. It's more high level than is ideal because we're testing the threading interaction as well as the specific buffering code. But it does give some sense of what's faster/slower and what a larger or smaller number of records.

GitOrigin-RevId: 93967dcbef5e9ceb7e57fa31df34be0f94968a10