Test isolate recreation on new client (#22570)
This PR adds tests for `IsolateWorker` to recreate the isolate when the client changes. I added `PauseClient` to both `BackendIsolateWorker` and `FunctionRunnerIsolateWorker`. It's redundant to test both because this logic is in the trait, but I think I'll add more tests that will use `PauseClient` later. This is pretty similar to how we test `AsyncLru` behavior that's in a separate thread.

GitOrigin-RevId: b52133ec9ab83b0bb0a9682555311a614386d78c