Set up funrun and aws lambda clients to receive streamed log lines (#22821)
The end goal is to stream log lines from actions. This is the client changes + all the wiring up of the sender, but nothing is actually sending log lines as they come in yet. The one thing that is set to surface log lines as they're streamed in are log streams (we can expand upon this later).

GitOrigin-RevId: cfaab4ac626b3ae00a28aa5ba292585ff75e2e6b