Backend + Funrun: Support transient transaction with existing writes (#22886)
This is the first change for enabling REPL-like experience: Providing a function source and immediately executing it, without any commits.

This PR enables existing writes on the transaction that's passed into function runner.

Such transaction is considered transient, and writes from the function run are not applied back to it (which doesn't matter in our REPL scenario since we won't commit the transaction anyway).

GitOrigin-RevId: 206a1fe39b0325a0ab8e6757131c0ef5e528c6f8