[transaction] run fetches in parallel (#23520)
using the same pattern as the rest of Transaction batching, we do a loop over the batch at the top of the function, and another loop at the end. But the middle is special:

use a `futures::join_all` to run the persistence fetches in parallel.

this is safe because `fetch_cache_misses` has no side effects -- it takes in the requested range and the cached values, it populates the cache misses with fetches, and returns results.

🥳

GitOrigin-RevId: bd872649f98ea98a9affa3559ec12efb31f5e121